% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilsInfoGrabbeR.R
\name{SoilsInfoGrabbeR}
\alias{SoilsInfoGrabbeR}
\title{SoilsSeeR Analysis Function}
\usage{
SoilsInfoGrabbeR(DSMW, ptz)
}
\arguments{
\item{DSMW}{character: Path to FAO-UNESCO DSMW}

\item{ptz}{vector: list of lat/long points, must be Deg. Min. Sec.!}
}
\description{
SoilsSeeR Analysis Function
}
